separacion del widget tw select file
